Continued moving logic to conversation class
